refactor(web): move blob storage configuration to a separate file
