+++ apps/web/src/blob-storages.ts
@@ -1,0 +1,31 @@
+import type { BlobStorage } from "./types";
+
+type StorageConfig = {
+  name?: string;
+  style: string;
+};
+
+export const STORAGE_CONFIGS: Record<BlobStorage, StorageConfig> = {
+  file_system: {
+    name: "File System",
+    style:
+      "bg-gray-100 hover:bg-gray-200 text-gray-800 hover:text-gray-900 dark:bg-gray-700 dark:text-gray-300 dark:hover:bg-gray-600 dark:hover:text-gray-200",
+  },
+  google: {
+    style:
+      "bg-slate-100 hover:bg-slate-200 text-slate-800 hover:text-slate-900 dark:bg-slate-700 dark:text-slate-300 dark:hover:bg-slate-600 dark:hover:text-slate-200",
+  },
+  swarm: {
+    style:
+      "bg-orange-100 hover:bg-orange-200 text-orange-800 hover:text-orange-900 dark:bg-orange-900 dark:text-orange-300 dark:hover:bg-orange-800 dark:hover:text-orange-200",
+  },
+  postgres: {
+    style:
+      "bg-blue-100 dark:bg-blue-900 text-blue-800 dark:text-blue-300 hover:bg-blue-200 dark:hover:bg-blue-800",
+  },
+  weavevm: {
+    name: "Load Network",
+    style:
+      "bg-gray-100 hover:bg-gray-200 text-gray-800 hover:text-gray-900 dark:bg-gray-700 dark:text-gray-300 dark:hover:bg-gray-600 dark:hover:text-gray-200",
+  },
+};
